# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IBKR REST Bridge API",
    "description": "RESTful API bridge for Interactive Brokers Client Portal Gateway",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic authentication for API access"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/account": {
      "get": {
        "tags": [
          "Account"
        ],
        "description": "Get account information including balances and positions",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string",
                      "example": "U1234567"
                    },
                    "accountType": {
                      "type": "string",
                      "example": "Individual"
                    },
                    "baseCurrency": {
                      "type": "string",
                      "example": "USD"
                    },
                    "balances": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "currency": {
                            "type": "string",
                            "example": "USD"
                          },
                          "cash": {
                            "type": "number",
                            "example": 10000
                          },
                          "totalValue": {
                            "type": "number",
                            "example": 50000
                          },
                          "buyingPower": {
                            "type": "number",
                            "example": 40000
                          }
                        },
                        "required": [
                          "currency",
                          "cash",
                          "totalValue",
                          "buyingPower"
                        ]
                      }
                    },
                    "positions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "conid": {
                            "type": "integer",
                            "example": 265598
                          },
                          "symbol": {
                            "type": "string",
                            "example": "AAPL"
                          },
                          "type": {
                            "type": "string",
                            "example": "STK"
                          },
                          "quantity": {
                            "type": "number",
                            "example": 100
                          },
                          "avgCost": {
                            "type": "number",
                            "example": 150.25
                          },
                          "marketValue": {
                            "type": "number",
                            "example": 17500
                          },
                          "unrealizedPnl": {
                            "type": "number",
                            "example": 475
                          }
                        },
                        "required": [
                          "conid",
                          "symbol",
                          "type",
                          "quantity",
                          "avgCost",
                          "marketValue",
                          "unrealizedPnl"
                        ]
                      }
                    }
                  },
                  "required": [
                    "accountId",
                    "accountType",
                    "baseCurrency",
                    "balances",
                    "positions"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/account/positions": {
      "get": {
        "tags": [
          "Account"
        ],
        "description": "Get all positions for the account",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "positions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "conid": {
                            "type": "integer",
                            "example": 265598
                          },
                          "symbol": {
                            "type": "string",
                            "example": "AAPL"
                          },
                          "type": {
                            "type": "string",
                            "example": "STK"
                          },
                          "quantity": {
                            "type": "number",
                            "example": 100
                          },
                          "avgCost": {
                            "type": "number",
                            "example": 150.25
                          },
                          "marketValue": {
                            "type": "number",
                            "example": 17500
                          },
                          "unrealizedPnl": {
                            "type": "number",
                            "example": 475
                          }
                        },
                        "required": [
                          "conid",
                          "symbol",
                          "type",
                          "quantity",
                          "avgCost",
                          "marketValue",
                          "unrealizedPnl"
                        ]
                      }
                    }
                  },
                  "required": [
                    "positions"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/instruments": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "description": "Search for instruments by symbol or name",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "AAPL"
            },
            "in": "query",
            "name": "q",
            "required": true,
            "description": "Search query (symbol or name)"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instruments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "conid": {
                            "type": "integer",
                            "example": 265598
                          },
                          "symbol": {
                            "type": "string",
                            "example": "AAPL"
                          },
                          "description": {
                            "type": "string",
                            "example": "APPLE INC"
                          },
                          "type": {
                            "type": "string",
                            "example": "STK"
                          },
                          "exchange": {
                            "type": "string",
                            "example": "NASDAQ"
                          }
                        },
                        "required": [
                          "conid",
                          "symbol",
                          "description",
                          "type",
                          "exchange"
                        ]
                      }
                    }
                  },
                  "required": [
                    "instruments"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quotes/{conid}": {
      "get": {
        "tags": [
          "Market Data"
        ],
        "description": "Get quote data for an instrument",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "265598"
            },
            "in": "path",
            "name": "conid",
            "required": true,
            "description": "Contract ID of the instrument"
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "conid": {
                      "type": "integer",
                      "example": 265598
                    },
                    "symbol": {
                      "type": "string",
                      "example": "AAPL"
                    },
                    "lastPrice": {
                      "type": "number",
                      "example": 175.5
                    },
                    "bidPrice": {
                      "type": "number",
                      "example": 175.45
                    },
                    "askPrice": {
                      "type": "number",
                      "example": 175.55
                    },
                    "bidSize": {
                      "type": "integer",
                      "example": 100
                    },
                    "askSize": {
                      "type": "integer",
                      "example": 200
                    },
                    "volume": {
                      "type": "integer",
                      "example": 5000000
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "conid",
                    "symbol",
                    "timestamp"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
}